+++ R/pkgstats-summary.R
@@ -53,7 +53,9 @@ pkgstats_summary <- function (s = NULL) {
     out$num_data_files <- s$data_stats [1]
     out$data_size_total <- s$data_stats [2]
     out$data_size_median <- s$data_stats [3]
-    out$translations <- paste0 (s$translations, collapse = ", ")
+
+    tr <- grepv ("^R\\-", basename (s$translations))
+    out$translations <- paste0 (gsub ("^R\\-", "", tr), collapse = ", ")
 
     out <- cbind (out, desc_summary (s$desc))